moved the components to separate files
