+++ src/components/NavigationMenu/NavigationMenu.tsx
@@ -1,98 +1,23 @@
-import * as Navigation from '@radix-ui/react-navigation-menu';
-import clsx from 'clsx';
-import { FC, KeyboardEvent, useRef } from 'react';
-import { Link, NavLink } from 'react-router-dom';
+import { List, Root } from '@radix-ui/react-navigation-menu';
+import { FC } from 'react';
 
 import style from './NavigationMenu.module.scss';
+import NavItem, { NavItemProps } from './NavItem';
 
 type NavigationMenuProps = {
   items: NavItemProps[];
 };
 
-type NavItemProps = {
-  label: string;
-  path: string;
-  dropdown?: NavItemProps[];
-};
-
-type ListItemProps = {
-  path: string;
-  label: string;
-  className?: string;
-  href: string;
-};
-
 const NavigationMenu: FC<NavigationMenuProps> = ({ items }) => {
   return (
-    <Navigation.Root className={style.menuRoot}>
-      <Navigation.List className={style.menuList}>
+    <Root className={style.menuRoot}>
+      <List className={style.menuList}>
         {items.map((item) => (
           <NavItem key={item.label} {...item} />
         ))}
-      </Navigation.List>
-    </Navigation.Root>
+      </List>
+    </Root>
   );
 };
-
-const NavItem: FC<NavItemProps> = ({ label, path, dropdown }) => {
-  const ref = useRef<HTMLAnchorElement>(null);
-
-  const handleKeyDown = (event: KeyboardEvent<HTMLButtonElement>) => {
-    if (event.key === 'Enter') {
-      ref.current?.focus();
-    }
-  };
-
-  return (
-    <Navigation.Item className={style.menuItem}>
-      <Navigation.Trigger className={style.menuTrigger} onKeyDown={handleKeyDown}>
-        <NavLink
-          to={path}
-          className={({ isActive }) =>
-            clsx(style.menuLink, {
-              [style.menuLinkActive]: isActive,
-            })
-          }
-          ref={ref}
-        >
-          {label}
-        </NavLink>
-      </Navigation.Trigger>
-      {dropdown && (
-        <Navigation.Content className={style.menuContent}>
-          <Navigation.Sub className={style.subMenu}>
-            <Navigation.List className={style.list}>
-              {dropdown.map((dropdownItem) => (
-                <ListItem
-                  key={dropdownItem.label}
-                  label={dropdownItem.label}
-                  path={dropdownItem.path}
-                  href={path}
-                />
-              ))}
-            </Navigation.List>
-          </Navigation.Sub>
-        </Navigation.Content>
-      )}
-    </Navigation.Item>
-  );
-};
-
-const ListItem: FC<ListItemProps> = ({ className, label, path, href }) => (
-  <li className={style.listItem}>
-    <Navigation.Link asChild>
-      <Link
-        to={`${href}${path}`}
-        className={clsx(style.listItemLink, className, {
-          [style.listItemLinkSeparator]:
-            label === 'Платные' || label === 'Облачные' || label === 'База игр',
-          [style.listItemLinkIcon]: label === 'Простые',
-        })}
-      >
-        {label}
-      </Link>
-    </Navigation.Link>
-  </li>
-);
 
 export default NavigationMenu;
